The PropCellTypes Class

The PropCellTypes class is a data storage class which contains information related to the cell type composition matrices used to simulate mixed transcriptional profiles. This matrix is stored in the prob.matrix slot while the rest contains additional information generated during the process and required in subsequent steps.

Details

See ?genMixedCellProp function for information about how cell type composition matrices are generated. Plots about the cell type proportion distributions can be accessed using the showProbPlot function (see ?showProbPlot for more details).

Slots

prob.matrix

Matrix of cell type proportions to simulate mixed transcriptional profiles.

cell.names

Matrix containing the name of cells to be used to generate the simulated mixed transcriptional profiles.

set.list

List of cells ordered according to the cell type they belong to.

set

Vector containing cell names present in the object.

method

Vector containing the method by which cell type proportions were generated.

plots

Plots showing the distribution of the cell type proportions. See ?showProbPlot for more details.

type.data

Character indicating the type of data contained: train or test.