The SpatialDDLS Class

The SpatialDDLSobject is the core of the SpatialDDLS package. This object stores different intermediate data needed for the construction of new deconvolution models, the spatial transcriptomics profiles to be deconvoluted, and the predicted cell type proportions.

Details

This object uses other classes to store different types of data generated during the workflow:

SingleCellExperiment class for single-cell RNA-Seq data storage, using sparse matrix from the Matrix package (dgCMatrix class) or HDF5Array class in case of using HDF5 files as back-end (see below for more information).

SpatialExperiment class for spatial transcriptomics data storage.

ZinbModel class with estimated parameters for the simulation of new single-cell profiles.

SummarizedExperiment class for simulated mixed transcriptional profiles storage.

PropCellTypes class for compositional cell type matrices. See ?PropCellTypes for details.

DeconvDLModel class to store the information related to deep neural network models. See ?DeconvDLModel for details.

In order to provide a way to work with large amounts of data on RAM-constrained machines, we provide the possibility to use HDF5 files as back-end to store count matrices of both real and simulated single-cell profiles by using the HDF5Array and DelayedArray classes from the homonymous packages.

Slots

single.cell.real

Real single-cell data stored in a SingleCellExperiment object. The count matrix is stored either as dgCMatrix or HDF5Array objects.

spatial.experiments

List of SpatialExperiment objects to be deconvoluted.

zinb.params

ZinbModel object with estimated parameters for the simulation of new single-cell expression profiles.

single.cell.simul

Simulated single-cell expression profiles using the ZINB-WaVE model.

prob.cell.types

PropCellTypes class with cell composition matrices built for the simulation of mixed transcriptional profiles with known cell composition.

mixed.profiles

List of simulated train and test mixed transcriptional profiles. Each entry is a SummarizedExperiment object. Count matrices can be stored as HDF5Array objects using HDF5 files as back-end in case of RAM limitations.

trained.model

DeconvDLModel object with information related to the trained model. See ?DeconvDLModel for more details.

deconv.spots

Deconvolution results. It consists of a list where each element corresponds to the results for each SpatialExperiment object contained in the spatial.experiments slot.

project

Name of the project.

version

Version of SpatialDDLS this object was built under.