Calculate evaluation metrics on test mixed transcriptional profiles

Calculate evaluation metrics on test mixed transcriptional profiles. By default, absolute error (AbsErr), proportional absolute error (ppAbsErr), squared error (SqrErr), and proportional squared error (ppSqrErr) are calculated for each test mixed profile. In addition, each of these metrics is aggregated according to three criteria: cell type (CellType), probability bins in ranges of 0.1 (pBin), and number of different cell types present in the spot (nCellTypes).

calculateEvalMetrics(object)

Arguments

object

SpatialDDLS object with a trained model in trained.model slot and the actual cell proportions of test mixed profiles in prob.cell.types slot.

Value

A SpatialDDLS object with is a DeconvDLModel object. The calculated metrics are stored in the test.deconv.metrics slot of the DeconvDLModel object.

See also